requirements.txt
sqlite3
hashlib
py-crypto
requests
fastapi
